Combine grids of analysis decisions into a single tibble

Combine grids of analysis decisions into a single tibble

Usage

combine_all_grids(
  .df,
  var_grid = NULL,
  filter_grid = NULL,
  model_grid = NULL,
  preprocessing = NULL,
  postprocessing = NULL
)

Arguments

.df

a data.frame representing the the original data

var_grid

a tibble created by create_var_grid. This grid should contain different versions of variables or sets of different versions of variables.

filter_grid

a tibble created by create_filter_grid. This grid contains all observation (row) filtering decisions.

model_grid

a tibble created by create_model_grid. This grid contains the actual model syntax (with possible alternatives) to run over the specified grids.

preprocessing

a tibble created by create_preprocess. It should contain code you want to run just after filtering but before analysis

postprocessing

a tibble created by create_postprocess. It should contain code that executes post hoc processing or analysis on the model object specified in the model_grid. This could be something like standardizing coefficients or running post hoc analysis (e.g., simple slopes).

Value

The output of this function will be a tibble of all combinations of analysis decisions. It contains the following columns:

decision

a numeric column from 1 to n combinations of decisions

filters

Filters (if they were included) as a list column.

variables

Variable combinations as a list column

model

The model, based on the decisions, that will be fit.

preprocess

any pre-processing steps contained in a list column

postprocess

any post-processing tasks contained in a list column

Examples

library(tidyverse)
library(multitool)
## Simulate some data
the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

## Create variable grid
my_var_grid <-
  create_var_grid(
    .df = the_data,
    iv  = c(iv1, iv2, iv3),
    mod = c(mod1, mod2, mod3),
    dv  = c(dv1, dv2),
    cov = c(cov1, cov2)
  )
#> cov variable group has 2 alternative variables
#> dv variable group has 2 alternative variables
#> iv variable group has 3 alternative variables
#> mod variable group has 3 alternative variables
#> 36 combinations (2*2*3*3 = 36)

## Create a filter grid
my_filter_grid <-
  create_filter_grid(
    .df = the_data,
    include1 == 0,
    include2 != 3,
    include2 != 2,
    scale(include3) > -2.5
  )
#> filters involving include1 has 2 alternative filtering criteria
#> filters involving include2 has 3 alternative filtering criteria
#> filters involving include3 has 2 alternative filtering criteria
#> 12 combinations (2*3*2 = 12)

## Create model grid
my_model_grid <-
  create_model_grid(
    lm({dv} ~ {iv} * {mod}),
    lm({dv} ~ {iv} * {mod} + {cov})
  )
#> Your model has 2 alternatives

## Add arbitrary code

# Code to execute after filtering step
my_preprocess <-
  create_preprocess(
    mutate({iv} := scale({iv}) |> as.numeric())
  )

# Code to execute after analysis is done
my_postprocess <-
  create_postprocess(
    aov()
  )

## Combine all grids together
my_full_grid <-
  combine_all_grids(
    the_data,
    my_var_grid,
    my_filter_grid,
    my_model_grid,
    my_preprocess,
    my_postprocess
  )
#> cov (variable) has 2 alternatives
#> dv (variable) has 2 alternatives
#> iv (variable) has 3 alternatives
#> mod (variable) has 3 alternatives
#> include1 (filter) has 2 alternatives
#> include2 (filter) has 3 alternatives
#> include3 (filter) has 2 alternatives
#> model has 2 alternatives
#> 864 combinations (2*2*3*3*2*3*2*2 = 864)