Create all combinations of filtering decisions for exploring exclusion criteria

Create all combinations of filtering decisions for exploring exclusion criteria

Usage

create_filter_grid(.df, ...)

Arguments

.df

the actual data as a data.frame for which you want to create filtering decisions

...

logical expressions to be used with filter separated by commas. Expressions should not be quoted.

Value

a list with four components:

summary

a summary tibble of your filters with the following columns: filter_expr (the filtering code), filter_group (which variable does the filter apply to), filtered_n (how many observations are there after the filter), and filter_type (does the expression filter or do nothing, which is auto-generated).

combinations

a summary of the number of alternatives per filtering decision

n_combinations

the total number of filtering combinations

grid

the actual filtering grid, the number of rows = n_combinations.

Examples

library(tidyverse)
library(multitool)

the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

my_filter_grid <-
  create_filter_grid(
    .df = the_data,
    include1 == 0,
    include2 != 3,
    include2 != 2,
    scale(include3) > -2.5
  )
#> filters involving include1 has 2 alternative filtering criteria
#> filters involving include2 has 3 alternative filtering criteria
#> filters involving include3 has 2 alternative filtering criteria
#> 12 combinations (2*3*2 = 12)

my_filter_grid
#> $summary
#> # A tibble: 7 × 4
#>   filter_expr                    filter_group filtered_n filter_type
#>   <chr>                          <chr>             <int> <chr>      
#> 1 include1 == 0                  include1            457 filter     
#> 2 include1 %in% unique(include1) include1            500 do nothing 
#> 3 include2 != 3                  include2            337 filter     
#> 4 include2 != 2                  include2            337 filter     
#> 5 include2 %in% unique(include2) include2            500 do nothing 
#> 6 scale(include3) > -2.5         include3            498 filter     
#> 7 include3 %in% unique(include3) include3            500 do nothing 
#> 
#> $combinations
#> # A tibble: 3 × 2
#>   filter_group n_alternatives
#>   <chr>                 <int>
#> 1 include1                  2
#> 2 include2                  3
#> 3 include3                  2
#> 
#> $n_combinations
#> [1] 12
#> 
#> $grid
#> # A tibble: 12 × 3
#>    include1                       include2                       include3       
#>    <chr>                          <chr>                          <chr>          
#>  1 include1 == 0                  include2 != 3                  scale(include3…
#>  2 include1 == 0                  include2 != 3                  include3 %in% …
#>  3 include1 == 0                  include2 != 2                  scale(include3…
#>  4 include1 == 0                  include2 != 2                  include3 %in% …
#>  5 include1 == 0                  include2 %in% unique(include2) scale(include3…
#>  6 include1 == 0                  include2 %in% unique(include2) include3 %in% …
#>  7 include1 %in% unique(include1) include2 != 3                  scale(include3…
#>  8 include1 %in% unique(include1) include2 != 3                  include3 %in% …
#>  9 include1 %in% unique(include1) include2 != 2                  scale(include3…
#> 10 include1 %in% unique(include1) include2 != 2                  include3 %in% …
#> 11 include1 %in% unique(include1) include2 %in% unique(include2) scale(include3…
#> 12 include1 %in% unique(include1) include2 %in% unique(include2) include3 %in% …
#>