Add arbitrary post hoc code to execute after each analysis

Add arbitrary post hoc code to execute after each analysis

Usage

create_postprocess(...)

Arguments

...

the literal code block (unquoted) you would like to execute after each analysis. If you have multiple post processing tasks, separate each distinct code chunk by a comma.

The code should be written to work with pipes (i.e., |> or %>%). Because the post hoc code comes last in each multiverse analysis step, the chosen model object will be passed to the post-processing code.

For example, if you fit a simple linear model like: lm(y ~ x1 + x2), and your post hoc code executes a call to anova, you would simply pass anova() to create_postprocess(). The underlying code would be:

data |> filters |> lm(y ~ x1 + x2, data = _) |> anova()

Value

a tibble with two columns, "postprocess" and "code". Each row representing a post processing task to execute. The "postprocess" column indicates the order (although its arbitrary) and the "code" column is the literal code to be ran.

You can add arguments to the code as you would normally. Always assume that the model you fit will passed as the first argument. However, for example, if you wanted to fit a simple slopes test with sim_slopes, you use glue

syntax. For example:

sim_slopes(pred = {iv}, modx = {mod})

The combine_all_grids will populate with the variables from your variable grid.

Examples

library(tidyverse)
library(multitool)