Early Life Adversity and Cognitive Ability

An unpublished experiment looking at the effect priming economic threat and childhood adversity on International Cognitive Ability Resource (ICAR) scores

Usage

icar_data

Format

A data frame with 413 rows and 63 variables:

sample

Lab or online subsample

condition

Experimental condition, 0 = control (computer crash); 1 = experimental (recession)

dems_gender

What gender do you identify with?

dems_age

What is your AGE in years? (e.g. 25)

dems_ethnicity

What is your ethnicity?

dems_edu

What is the highest level of education you have completed?

dems_us_born

Did you grow up in the United States?

dems_english_native

Did you grow up speaking English?

dems_fluency

How fluent are you with the English language?

dems_lang

Is English the language you know best?

child_unp_obj1

How many times did your parents or legal guardians change jobs or occupational status?

child_unp_obj2

How many times did you move or change residence (move to a different house, neighborhood, city, or state)?

child_unp_obj3

How many times were there changes in your familial circumstances? (divorce, parents starting new relationships, parents leaving the home)

child_unp_changes1

Economic status:

child_unp_changes2

Family environment:

child_unp_changes3

Your childhood neighborhood environment:

child_unp_changes4

Your childhood school environment:

child_unp_subj1

My family life was generally inconsistent and unpredictable from day-to-day.

child_unp_subj2

My parent(s) frequently had arguments or fights with each other or other people in my childhood.

child_unp_subj3

My parents had a difficult divorce or separation during this time.

child_unp_subj4

People often moved in and out of my house on a pretty random basis.

child_unp_subj5

When I woke up, I often didn't know what could happen in my house that day.

child_unp_subj6

My family environment was often tense and "on edge".

child_unp_subj7

Things were often chaotic in my house.

child_unp_subj8

I had a hard time knowing what my parent(s) or other people in my house were going to say.

child_ses_subj1

I grew up in a relatively wealthy neighborhood.

child_ses_subj2

I felt relatively wealthy compared to the other kids.

child_ses_subj3

My family usually had enough money for things when I was growing up.

child_income

What was your household income when you were growing up?

unp_obj_mean

Childhood unpredictability (objective) mean score

unp_changes_mean

Childhood unpredictability (changes) mean score

unp_subj_mean

Childhood unpredictability (subjective) mean score

ses_subj_mean

Childhood socioeconomic status (objective) mean score

vr_04

Verbal reasoning item 4

vr_17

Verbal reasoning item 17

ln_07

Letter-number item 7

ln_58

Letter-number item 58

mx_45

Matrix reasoning item 45

mx_46

Matrix reasoning item 46

r3d_06

Mental Rotation item 6

r3d_08

Mental Rotation item 8

vr_16

Verbal reasoning item 16

vr_19

Verbal reasoning item 19

ln_33

Letter-number item 33

ln_34

Letter-number item 34

mx_47

Matrix reasoning item 47

mx_55

Matrix reasoning item 55

r3d_03

Mental Rotation item 3

r3d_04

Mental Rotation item 4

ln_sum

Total score (out of 4) on the letter-number items

mx_sum

Total score (out of 4) on the matrix reasoning items

vr_sum

Total score (out of 4) on the verbal reasoning items

r3d_sum

Total score (out of 4) on the 3d rotation items

icar_sum

Total score (out of 16) on the ICAR battery

time_condition

Time in seconds the experimental manipulation displayed (should be about 60 seconds)

time_icar

Time in seconds it took to complete the whole ICAR battery

time_ln

Time in seconds it took to complete the lettter-number items

time_mx

Time in seconds it took to complete the matrix reasoning items

time_vr

Time in seconds it took to complete the verbal reasoning items

time_r3d

Time in seconds it took to complete the 3d rotation items

att_interrupt

Were you ever interrupted by anyone or anything during the time you spent on this survey?

att_one_sitting

Did you complete this survey in one sitting?

att_getup

Did you ever get up and leave your computer during the study (for any reason)?

Source

data_URL