Show multiverse data code pipelines

Each show_code* function should be self-explanatory - they indicate where along the multiverse pipeline to extract code. The goal of these functions is to create a window into each multiverse decision set context/results and allow the user to inspect specific decisions straight from the code that produced it.

Usage

show_code_subgroups(
  .grid,
  decision_num,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

show_code_filter(
  .grid,
  decision_num,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

show_code_preprocess(
  .grid,
  decision_num,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

show_code_model(
  .grid,
  decision_num,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

show_code_postprocess(
  .grid,
  decision_num,
  post_step = 1,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

show_code_summary_stats(
  .grid,
  decision_num,
  summary_set = 1,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

show_code_corrs(
  .grid,
  decision_num,
  corr_set = 1,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

show_code_reliabilities(
  .grid,
  decision_num,
  rel_set = 1,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

Arguments

.grid

a full decision grid created by expand_decisions.

decision_num

numeric. Indicates which 'universe' in the multiverse to show underlying code.

copy

logical. Whether to copy the pipeline code to the clipboard using write_clip. Defaults to FALSE.

console

logical. Whether to send the code to the console in RStudio. Defaults to TRUE but requires that the code be running in RStudio.

execute

logical. If sending to the console, whether to immediately run the code in the console. Defaults to FALSE.

...

additional arguments passed to rstudioapi::sendToConsole()

post_step

numeric. For show_code_postprocess, Which post-processing step to print. Default is set to the 1.

summary_set

numeric. For show_code_summary_stats, Which set of summary statistics to print. Default is set to the 1.

corr_set

numeric. For show_code_corrs, Which set of correlations to print. Default is set to the 1.

rel_set

numeric. For show_code_reliabilities, Which set of reliabilities to print. Default is set to the 1.

Value

the code that generated results up to the specified point in an analysis pipeline. The code is printed in the console and can be optionally copied to the clipboard.

Functions

show_code_filter(): Show the code up to the filtering stage

show_code_preprocess(): Show the code up to the preprocessing stage

show_code_model(): Show the code up to the modeling stage

show_code_postprocess(): Show the code up to the post-processing stage

show_code_summary_stats(): Show the code for computing summary statistics

show_code_corrs(): Show the code for computing correlations

show_code_reliabilities(): Show the code for computing scale reliability